Update made for heroku deployment
