+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/siesgstarena/epicentre
+
+go 1.14
+
+require (
+	github.com/caarlos0/env/v6 v6.2.1
+	github.com/gin-gonic/gin v1.6.2
+	go.uber.org/zap v1.14.1
+	gopkg.in/natefinch/lumberjack.v2 v2.0.0
+)
