--- src/go.mod
@@ -1,10 +1,0 @@
-module github.com/siesgstarena/epicentre
-
-go 1.14
-
-require (
-	github.com/caarlos0/env/v6 v6.2.1 // indirect
-	github.com/gin-gonic/gin v1.6.2
-	go.uber.org/zap v1.14.1 // indirect
-	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
-)
